% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chplot.R
\name{chplot}
\alias{chplot}
\title{chplot}
\usage{
chplot(sequences, names)
}
\arguments{
\item{sequences}{Character vector of sequences}

\item{names}{Character vector of protein names}
}
\description{
Creates a charge-hydropathy plot for multiple proteins
}
\examples{

s <- c('MGNHAGKRELNAEKASTNSETNRGESEKKRNLGELSRTTSEDNEVFGEADANQNNGTSSQ
DTAVTDSKRTADPKNAWQDAHPADPGSRPHLIRLFSRDAPGREDNTFKDRPSESDELQTI
QEDSAATSESLDVMASQKRPSQRHGSKYLATASTMDHARHGFLPRHRDTGILDSIGRFFG
GDRGAPKRGSGKDSHHPARTAHYGSLPQKSHGRTQDENPVVHFFKNIVTPRTPPPSQGKG
RGLSLSRFSWGAEGQRPGFGYGGRASDYKSAHKGFKGVDAQGTLSKIFKLGGRDSRSGSP
MARR')

n <- c('MBP P02686')

chplot(s,n)
}
