% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprof.R
\name{cprof}
\alias{cprof}
\title{cprof}
\usage{
cprof(sequences, names)
}
\arguments{
\item{sequences}{Character vector of sequences}

\item{names}{Character vector of protein names}
}
\description{
Plots the residue profile based on reference data sets
}
\examples{

s <- c('MGNHAGKRELNAEKASTNSETNRGESEKKRNLGELSRTTSEDNEVFGEADANQNNGTSSQ
DTAVTDSKRTADPKNAWQDAHPADPGSRPHLIRLFSRDAPGREDNTFKDRPSESDELQTI
QEDSAATSESLDVMASQKRPSQRHGSKYLATASTMDHARHGFLPRHRDTGILDSIGRFFG
GDRGAPKRGSGKDSHHPARTAHYGSLPQKSHGRTQDENPVVHFFKNIVTPRTPPPSQGKG
RGLSLSRFSWGAEGQRPGFGYGGRASDYKSAHKGFKGVDAQGTLSKIFKLGGRDSRSGSP
MARR')

n <- c('MBP P02686')

cprof(s,n)
}
