% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_CDF.R
\name{delta_CDF}
\alias{delta_CDF}
\title{Delta Cumulative Distribution Frequency}
\usage{
delta_CDF(CDF)
}
\arguments{
\item{CDF}{List of CDF values for each protein}
}
\value{

}
\description{
Delta Cumulative Distribution Frequency
}
\examples{
\dontrun{
d_CDF <- delta_CDF(CDF)
}
}
