% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_CH.R
\name{delta_CH}
\alias{delta_CH}
\title{Delta Charge Hydropathy}
\usage{
delta_CH(sequences, names)
}
\arguments{
\item{sequences}{Character vector of sequences, do not use FASTA format}

\item{names}{Character vector of protein names}
}
\value{

}
\description{
Delta Charge Hydropathy
}
\examples{
\dontrun{
d_CH <- delta_CH(PROTEINS$SEQUENCE)
}
}
