% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CDF.R
\name{get_CDF}
\alias{get_CDF}
\title{Calculate Cumulative Distribution Frequency}
\usage{
get_CDF(PONDR_VLXT, ID)
}
\arguments{
\item{PONDR_VLXT}{List of disorder scores for each protein}

\item{ID}{Chracter vector of protein names}
}
\value{

}
\description{
Calculate Cumulative Distribution Frequency
}
\examples{
\dontrun{
CDF <- get_CDF(PONDR_VLXT, PROTEINS$ID)
}
}
