% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq.R
\name{get_seq}
\alias{get_seq}
\title{Get sequences from UniProt, also stores them in FASTA format}
\usage{
get_seq(ID)
}
\arguments{
\item{ID}{Character vector of UniProt IDs}
}
\value{

}
\description{
Get sequences from UniProt, also stores them in FASTA format
}
\examples{
\dontrun{
PROTEINS <- get_seq(ID)
}
}
