% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CDF.R
\name{plot_CDF}
\alias{plot_CDF}
\title{Create CDF plot for multiple proteins}
\usage{
plot_CDF(CDF)
}
\arguments{
\item{CDF}{List of CDF values}
}
\value{

}
\description{
Create CDF plot for multiple proteins
}
\examples{
\dontrun{
plot_CDF(CDF)
}
}
