% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CH.R
\name{plot_CH}
\alias{plot_CH}
\title{Create a charge-hydropathy plot for multiple proteins}
\usage{
plot_CH(sequences, ID)
}
\arguments{
\item{sequences}{Character vector of sequences}

\item{ID}{Character vector of protein names}
}
\value{

}
\description{
Create a charge-hydropathy plot for multiple proteins
}
\examples{
\dontrun{
plot_CH(PROTEINS$SEQUENCE,PROTEINS$ID)
}
}
